Move `JsonRpcResponseType` to `RpcResponseType` in the plugin API module
