+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/handlers/JsonRpcArrayExecutor.java
@@ -22,8 +22,8 @@ import org.hyperledger.besu.ethereum.api.jsonrpc.context.ContextKey;
 import org.hyperledger.besu.ethereum.api.jsonrpc.execution.JsonRpcExecutor;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcErrorResponse;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcResponse;
-import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcResponseType;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.RpcErrorType;
+import org.hyperledger.besu.plugin.services.rpc.RpcResponseType;
 
 import java.io.IOException;
 
@@ -74,7 +74,7 @@ public class JsonRpcArrayExecutor extends AbstractJsonRpcExecutor {
       generator.writeStartArray();
       for (int i = 0; i < rpcRequestBatch.size(); i++) {
         JsonRpcResponse response = processMaybeRequest(rpcRequestBatch.getValue(i));
-        if (response.getType() != JsonRpcResponseType.NONE) {
+        if (response.getType() != RpcResponseType.NONE) {
           generator.writeObject(response);
         }
       }
