+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/response/JsonRpcSuccessResponse.java
@@ -14,6 +14,8 @@
  */
 package org.hyperledger.besu.ethereum.api.jsonrpc.internal.response;
 
+import org.hyperledger.besu.plugin.services.rpc.RpcResponseType;
+
 import java.util.Objects;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
@@ -50,8 +52,8 @@ public class JsonRpcSuccessResponse implements JsonRpcResponse {
 
   @Override
   @JsonIgnore
-  public JsonRpcResponseType getType() {
-    return JsonRpcResponseType.SUCCESS;
+  public RpcResponseType getType() {
+    return RpcResponseType.SUCCESS;
   }
 
   @Override
