+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/response/JsonRpcUnauthorizedResponse.java
@@ -14,6 +14,8 @@
  */
 package org.hyperledger.besu.ethereum.api.jsonrpc.internal.response;
 
+import org.hyperledger.besu.plugin.services.rpc.RpcResponseType;
+
 import java.util.Arrays;
 import java.util.Objects;
 
@@ -50,8 +52,8 @@ public class JsonRpcUnauthorizedResponse implements JsonRpcResponse {
 
   @Override
   @JsonIgnore
-  public JsonRpcResponseType getType() {
-    return JsonRpcResponseType.UNAUTHORIZED;
+  public RpcResponseType getType() {
+    return RpcResponseType.UNAUTHORIZED;
   }
 
   @Override
