+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/ipc/JsonRpcIpcService.java
@@ -20,8 +20,8 @@ import org.hyperledger.besu.ethereum.api.jsonrpc.execution.JsonRpcExecutor;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.JsonRpcRequest;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcErrorResponse;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcResponse;
-import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcResponseType;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.RpcErrorType;
+import org.hyperledger.besu.plugin.services.rpc.RpcResponseType;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -155,7 +155,7 @@ public class JsonRpcIpcService {
                                                 .filter(
                                                     jsonRpcResponse ->
                                                         jsonRpcResponse.getType()
-                                                            != JsonRpcResponseType.NONE)
+                                                            != RpcResponseType.NONE)
                                                 .toArray(JsonRpcResponse[]::new);
 
                                         socket.write(
