+++ ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/privacy/methods/MultiTenancyRpcMethodDecoratorTest.java
@@ -22,10 +22,10 @@ import org.hyperledger.besu.ethereum.api.jsonrpc.internal.JsonRpcRequestContext;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.methods.JsonRpcMethod;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcErrorResponse;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcResponse;
-import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcResponseType;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcSuccessResponse;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcUnauthorizedResponse;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.RpcErrorType;
+import org.hyperledger.besu.plugin.services.rpc.RpcResponseType;
 
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.auth.User;
@@ -58,7 +58,7 @@ public class MultiTenancyRpcMethodDecoratorTest {
     assertThat(tokenRpcDecorator.getName()).isEqualTo("delegate");
 
     final JsonRpcResponse response = tokenRpcDecorator.response(rpcRequestContext);
-    assertThat(response.getType()).isEqualTo(JsonRpcResponseType.SUCCESS);
+    assertThat(response.getType()).isEqualTo(RpcResponseType.SUCCESS);
     final JsonRpcSuccessResponse successResponse = (JsonRpcSuccessResponse) response;
     assertThat(successResponse.getResult()).isEqualTo("b");
   }
@@ -73,7 +73,7 @@ public class MultiTenancyRpcMethodDecoratorTest {
     assertThat(tokenRpcDecorator.getName()).isEqualTo("delegate");
 
     final JsonRpcResponse response = tokenRpcDecorator.response(rpcRequestContext);
-    assertThat(response.getType()).isEqualTo(JsonRpcResponseType.UNAUTHORIZED);
+    assertThat(response.getType()).isEqualTo(RpcResponseType.UNAUTHORIZED);
     final JsonRpcUnauthorizedResponse errorResponse = (JsonRpcUnauthorizedResponse) response;
     assertThat(errorResponse.getErrorType()).isEqualTo(RpcErrorType.UNAUTHORIZED);
   }
@@ -89,7 +89,7 @@ public class MultiTenancyRpcMethodDecoratorTest {
     assertThat(tokenRpcDecorator.getName()).isEqualTo("delegate");
 
     final JsonRpcResponse response = tokenRpcDecorator.response(rpcRequestContext);
-    assertThat(response.getType()).isEqualTo(JsonRpcResponseType.ERROR);
+    assertThat(response.getType()).isEqualTo(RpcResponseType.ERROR);
     final JsonRpcErrorResponse errorResponse = (JsonRpcErrorResponse) response;
     assertThat(errorResponse.getErrorType()).isEqualTo(RpcErrorType.INVALID_REQUEST);
   }
